requirements.txt
absl-py==1.4.0
agora-token-builder==1.0.0
aioredis==1.3.1
amqp==5.1.1
appdirs==1.4.4
asgiref==3.4.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
bandit==1.7.4
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.0
bootstrap4==0.1.0
cachetools==5.3.0
celery==5.2.7
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==4.0.0
charset-normalizer==2.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
comm==0.1.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2022.1
cryptography==3.4.7
cycler==0.11.0
daphne==3.0.2
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.0.7
django-allauth==0.45.0
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-crispy-forms==1.12.0
django-extensions==3.2.1
django-jasmine==0.5.1
django-js-asset==1.2.2
djangorestframework==3.14.0
docopt==0.6.2
drf-yasg==1.21.5
entrypoints==0.4
executing==1.2.0
factory-boy==3.2.1
Faker==15.3.4
fastjsonschema==2.15.3
filelock==3.8.0
flatbuffers==23.1.21
fonttools==4.33.3
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.51.1
h5py==3.8.0
hiredis==2.2.2
hyperlink==21.0.0
idna==3.2
imageio==2.25.1
importlib-metadata==4.12.0
incremental==21.3.0
inflection==0.5.1
install==1.3.5
ipykernel==6.19.4
ipython==8.7.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
keras==2.11.0
kiwisolver==1.4.3
kombu==5.2.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mistune==0.8.4
msgpack==1.0.5
msgpack-python==0.5.6
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.0
nltk==3.7
notebook-as-pdf==0.5.0
numpy==1.23.0
oauthlib==3.1.1
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pbr==5.10.0
pi==0.1.2
pickleshare==0.7.5
Pillow==9.0.1
pipreqs==0.4.11
platformdirs==2.5.4
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
psycopg2==2.9.1
psycopg2-binary==2.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
pyee==8.2.2
Pygments==2.12.0
PyJWT==2.4.0
pyOpenSSL==20.0.1
pyparsing==3.0.9
PyPDF2==2.4.0
pyppeteer==1.0.2
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
PyWavelets==1.4.1
pywin32==304
PyYAML==6.0
pyzmq==23.2.0
redis==4.5.1
regex==2022.6.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.8.1
seaborn==0.11.2
service-identity==21.1.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.6.2
stevedore==4.0.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.2.3
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.0
traitlets==5.8.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.6
vine==5.0.0
virtualenv==20.16.7
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
Werkzeug==2.2.3
wordcloud==1.8.1
wrapt==1.14.1
yarg==0.1.9
zipp==3.8.0
zope.interface==5.4.0
